+++ 5-AccessControl/1-call-api-roles/API/package.json
@@ -14,7 +14,7 @@
     "lowdb": "^1.0.0",
     "morgan": "^1.10.0",
     "passport": "^0.6.0",
-    "passport-azure-ad": "^4.3.3"
+    "passport-azure-ad": "^4.3.4"
   },
   "main": "app.js",
   "repository": {
